Prepare tests to use environments with group tags
The idea is to add a `groups` field to the osprey target. This group
field would contain a list of `tags` or `group` names to which an
environment belongs to.

The user then can login to all targets belonging to a group by providing
a tag, or list of tags.

Additional commands can be added to list groups, targets, and targets by
group so that the user can easily see what will osprey try to reach when
loging in.