Add groups to osprey configuration
This feature allows a user to specify groups for each target and to
specify a default group to log in to.

The user, login and logout commands now allow a `--group` flag to
select the targets to login to.

If the `group` flag is provided, osprey will filter the targets and
execute the command against the ones with a matching group.
If not provided opsrey will look for a `default-group` defined in
the config. If it exists, it will use that group to filter the targets.
If no default exists, then it will execute the command against all
the targets without a group.

Note that adding a default-group will make it impossible to log in to
targets without a group.