Add ospreyconfig validation for shadowed targets
A target can be shadowed if an osprey configuration defines a
default group, but fails to provide a group for every target.
If this were to happen, because the default group will make it
impossible for the osprey config to reach the ungrouped targets, the new
validation will fail any command, and display the appropriate error.