Fix latest release publishing
Treat `latest` as a version in bintray, and deploy the current version
and the latest version artifacts separately.

The bintray deploy task originally created included the latest artifacts
as part of the bintray deploy for a specific version.
Because of this, subsequent releases failed to update the artifacts, as
they were tied to a specific version, e.g.
```
Bintray Upload] Bintray response: 409 Conflict. Unable to upload files:
An artifact with the path '[...]' already exists under another version.
```