force the e2e test dex server to listen on both ipv4 and 6
On some machines (mine...) `localhost` could resolve to an ipv6
address.  That borks the test.  Listening on `127.0.0.1` might seem like
itʼs specific to ipv4 only, but thatʼs an illusion.  (I am pretty sure
that) that actually listens on both the ipv4 address as well as the
equivalent ipv6 address (`:ffff:127.0.0.1`).