Allow osprey client to fetch CA cert from API Server
Add an optional "api-server" field to the Osprey client config file.
If this is used, the API Server CA certificate will be fetched directly
from the API Server rather than relying on a deployment of Osprey server
to serve the CA.

To use this, the cluster must have a ConfigMap called kube-root-ca.crt
in the kube-public namespace that is accessible to anonymous users.
The following should work:

curl -k https://APISERVER/api/v1/namespaces/kube-public/configmaps/kube-root-ca.crt

This ConfigMap is created automatically as part of the RootCAConfigMap
feature gate introduced in Kubernetes 1.13, which became enabled by
default in v1.20.